test: add tests for uncovered functions and fix sample/shuffle bug
- Add tests for builtin functions: copy, clone, equals, flatten, defaults,
  genOtpCode, checkErr, checkEmpty, typeOf with detailed mode
- Add tests for CompiledRegex type methods
- Add tests for env module: get/set, getOr, has, all, map, expand, cwd,
  pid, ppid, exe, streams, lookup, getInt, getBool, mixArgs
- Add tests for log module: debug, info, warn, error, format, isLevel,
  print, printNoNL, printf, withPrefix, json, stack
- Add tests for collections module: union, intersection, difference,
  sample, shuffle, chunk, flattenDeep, groupBy, countBy, partition
- Add tests for debug module: type, typeTag, inspect, memStats, gc,
  goroutines, stack, callers, assert, assertEquals, dump, isNull,
  isTruthy, reflect, nanoTime, elapsed
- Add tests for sort module: numbers, numbersDesc, strings, stringsDesc,
  reverse, isSorted, min, max, by

Fix integer overflow bug in sample/shuffle functions that could cause
negative array indices.

+++ pkg/stdlib/collections.go
@@ -465,7 +465,7 @@ func init() {
 				result := make([]objects.Object, n)
 				copy(result, arr.Elements)
 				for i := n - 1; i > 0; i-- {
-					j := int(simpleRand()) % (i + 1)
+					j := int(simpleRand() % uint64(i+1))
 					result[i], result[j] = result[j], result[i]
 				}
 				return Array(result...)
@@ -491,13 +491,13 @@ func init() {
 					}
 				}
 				if count == 1 {
-					idx := int(simpleRand()) % len(arr.Elements)
+					idx := int(simpleRand() % uint64(len(arr.Elements)))
 					return arr.Elements[idx]
 				}
 				// Return multiple samples
 				result := []objects.Object{}
 				for i := int64(0); i < count && i < int64(len(arr.Elements)); i++ {
-					idx := int(simpleRand()) % len(arr.Elements)
+					idx := int(simpleRand() % uint64(len(arr.Elements)))
 					result = append(result, arr.Elements[idx])
 				}
 				return Array(result...)
